requirements.txt
flask
flask-cors
sentence-transformers
torch
python-Levenshtein
rapidfuzz
textblob
pyspellchecker
gunicorn
scipy
